# openapi/LupiraMtgApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira MTG API",
    "description": "Backend for the Lupira MTG mobile app — Magic: The Gathering card metadata, scan-based recognition, and per-user collection management. Authenticate with a Bearer token issued by Authentik (OIDC).",
    "version": "v1"
  },
  "paths": {
...
    "/collections": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "List the caller's collections.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create a new collection.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get a collection with its cards.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Collections"
        ],
        "summary": "Rename a collection.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Soft-delete a collection.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "List the cards in a collection.",
        "description": "Sorted by card name (ascending). `take` 1–200 (default 50). `skip` for paging.\nResponse includes `total`.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInstanceListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Add a card to a collection (manual, no scan).",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCardToCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInstanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards/{instanceId}": {
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Remove a card from a collection.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards/{instanceId}/move": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Move a card to another collection.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInstanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards/bulk": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Add many cards to a collection in one call.",
        "description": "Body: `{ items: [{ printingId, isFoil?, language?, condition?, count? }] }`.\nPer-item `count` is clamped to 1..50; total instances per call capped at 500.\nReturns the newly-created `CardInstanceResponse` rows in the same order they\nwere generated. Validates every `printingId` up front — if any is unknown the\nwhole call is rejected with `400`, no partial writes.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAddCardsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkAddCardsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards/bulk-delete": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Remove many cards from a collection in one call.",
        "description": "Body: `{ instanceIds: [...] }`. Idempotent on missing ids — returns\n`{ removedCount, missingCount }` so the client can tell the user what landed.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRemoveCardsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkRemoveCardsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/collections/{collectionId}/cards/bulk-move": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Move many cards to another collection in one call.",
        "description": "Body: `{ instanceIds: [...], toCollectionId }`. Both source and destination\nmust exist and belong to the caller. Returns the moved instances (now with\nthe destination collection set) plus a count of ids that weren't in the source.",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMoveCardsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkMoveCardsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCardToCollectionRequest": {
        "required": [
          "printingId"
        ],
        "type": "object",
        "properties": {
          "printingId": {
            "type": "string"
          },
          "isFoil": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          }
        }
      },
...
      "BulkAddCardItem": {
        "required": [
          "printingId"
        ],
        "type": "object",
        "properties": {
          "printingId": {
            "type": "string"
          },
          "isFoil": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "BulkAddCardsRequest": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkAddCardItem"
            }
          }
        }
      },
      "BulkAddCardsResponse": {
        "required": [
          "added"
        ],
        "type": "object",
        "properties": {
          "added": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardInstanceResponse"
            }
          }
        }
      },
      "BulkMoveCardsRequest": {
        "required": [
          "instanceIds",
          "toCollectionId"
        ],
        "type": "object",
        "properties": {
          "instanceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "toCollectionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BulkMoveCardsResponse": {
        "required": [
          "moved",
          "missingCount"
        ],
        "type": "object",
        "properties": {
          "moved": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardInstanceResponse"
            }
          },
          "missingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "BulkRemoveCardsRequest": {
        "required": [
          "instanceIds"
        ],
        "type": "object",
        "properties": {
          "instanceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "BulkRemoveCardsResponse": {
        "required": [
          "removedCount",
          "missingCount"
        ],
        "type": "object",
        "properties": {
          "removedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "missingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CardImageUrls": {
        "type": "object",
        "properties": {
          "normal": {
            "type": [
              "null",
              "string"
            ]
          },
          "artCrop": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CardInstanceListResponse": {
        "required": [
          "cards",
          "total"
        ],
        "type": "object",
        "properties": {
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardInstanceResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CardInstanceResponse": {
        "required": [
          "instanceId",
          "printing",
          "isFoil",
          "language",
          "condition",
          "acquiredAt"
        ],
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string",
            "format": "uuid"
          },
          "printing": {
            "$ref": "#/components/schemas/CardPrintingResponse"
          },
          "isFoil": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "acquiredAt": {
            "type": "string",
            "format": "date-time"
          },
          "collectionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "collectionName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CardPrintingResponse": {
        "required": [
          "id",
          "oracleId",
          "name",
          "setCode",
          "setName",
          "collectorNumber",
          "colorIdentity",
          "rarity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "oracleId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "setCode": {
            "type": "string"
          },
          "setName": {
            "type": "string"
          },
          "collectorNumber": {
            "type": "string"
          },
          "colorIdentity": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rarity": {
            "type": "string"
          },
          "manaCost": {
            "type": [
              "null",
              "string"
            ]
          },
          "cmc": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "float"
          },
          "images": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardImageUrls"
              }
            ]
          },
          "prices": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          "faces": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CardFaceResponse"
            }
          }
        }
      },
...
      "CollectionDetailResponse": {
        "required": [
          "id",
          "name",
          "cards",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardInstanceResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CollectionListResponse": {
        "required": [
          "collections"
        ],
        "type": "object",
        "properties": {
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionResponse"
            }
          }
        }
      },
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "cardCount",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cardCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateCollectionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "MoveCardRequest": {
        "required": [
          "toCollectionId"
        ],
        "type": "object",
        "properties": {
          "toCollectionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RenameCollectionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PoC device token from `POST /me/register`. Send as `Authorization: Bearer lmtg_<token>`. Replaced by Authentik OIDC tokens in the future Path C migration (see plan).",
        "scheme": "bearer"
      }
    }
  },
...
}